j ai travailler sur la methode create transaction

+++ example-app/database/migrations/2025_03_10_104318_create_wallets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->string('balance');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->onUpdate('cascade');
             $table->timestamps();
